Aprimoramento da geração automática das imagens

+++ app/Http/Controllers/Api/ProductController.php
@@ -151,40 +151,46 @@ class ProductController extends Controller
         //
     }
 
-    public function generateImages()
+    public function generateImages(Request $request)
     {
-        $products = DB::table('products')
-            ->select('base_code', 'old_code', 'product_code', 'id')
-            ->get();
+        if ($request->input('key') === 'generate-images') {
 
-        DB::table('product_images')->truncate();
+            $products = DB::table('products')
+                ->select('base_code', 'old_code', 'product_code', 'id')
+                ->get();
 
-        $real_path = realpath('');
+            DB::table('product_images')->truncate();
 
-        foreach ($products as $product) {
+            $real_path = realpath('');
 
-            if (! is_dir("{$real_path}/img/products/{$product->base_code}")) {
-                mkdir("{$real_path}/img/products/{$product->base_code}", 0777, true);
+            foreach ($products as $product) {
+
+                if (!is_dir("{$real_path}/img/products/{$product->base_code}")) {
+                    mkdir("{$real_path}/img/products/{$product->base_code}", 0777, true);
+                }
+
+                $product_image = glob("{$real_path}/img/products/{$product->product_code}.jpg");
+
+                if (count($product_image) > 0) {
+
+                    copy($product_image[0],
+                        "{$real_path}/img/products/{$product->base_code}/{$product->product_code}.jpg");
+
+                    DB::table('product_images')
+                        ->insert([
+                            'product_id' => $product->id,
+                            'image_name' => $product->product_code . '.jpg',
+                            'image_url' => "https://api.marcalaser.com/img/products/{$product->base_code}/{$product->product_code}.jpg",
+                            'is_default' => 1,
+                            'created_at' => date('Y-m-d H:i:s'),
+                            'updated_at' => date('Y-m-d H:i:s'),
+                        ]);
+                }
             }
 
-            $product_image = glob("{$real_path}/img/products/{$product->product_code}.jpg");
-
-            if (count($product_image) > 0) {
-
-                copy($product_image[0], "{$real_path}/img/products/{$product->base_code}/{$product->product_code}.jpg");
-
-                DB::table('product_images')
-                    ->insert([
-                        'product_id' => $product->id,
-                        'image_name' => $product->product_code . '.jpg',
-                        'image_url' => "https://api.marcalaser.com/img/products/{$product->base_code}/{$product->product_code}.jpg",
-                        'is_default' => 1,
-                        'created_at' => date('Y-m-d H:i:s'),
-                        'updated_at' => date('Y-m-d H:i:s'),
-                    ]);
-            }
+            return response()->json('Imagens geradas com sucesso!');
+        } else {
+            die('Acesso negado!');
         }
-
-        return response()->json('Imagens geradas com sucesso!');
     }
 }
